add url support to select server
