+++ src/context/AuthContext.js
@@ -5,6 +5,7 @@ import {
   BASE_URL,
   CLOSE,
   HTTP_GET_MAX_WAIT_TIME,
+  SERVER_KEY,
   TOKEN_ID,
 } from '@common/constants'
 import {
@@ -13,12 +14,14 @@ import {
   unAuthenticated,
 } from '@utils/network'
 import NetworkErrorPopup from '@components/NetworkErrorPopUp'
+import useLocalStorage from '@hooks/useLocalStorage'
 
 export const AuthContext = createContext({})
 
 export default function AuthContextProvider(props) {
   const [authentication, setAuthentication] = useState(null)
   const [networkError, setNetworkError] = useState(null)
+  const [server, setServer] = useLocalStorage(SERVER_KEY, BASE_URL)
 
   /**
    * in the case of a network error, process and display error dialog
@@ -38,7 +41,7 @@ export default function AuthContextProvider(props) {
     const auth = await myAuthStore.getItem('authentication')
 
     if (auth) { // verify that auth is still valid
-      doFetch('https://git.door43.org/api/v1/user', auth, HTTP_GET_MAX_WAIT_TIME)
+      doFetch(`${server}/api/v1/user`, auth, HTTP_GET_MAX_WAIT_TIME)
         .then(response => {
           const httpCode = response?.status || 0
 
@@ -94,10 +97,12 @@ export default function AuthContextProvider(props) {
     state: {
       authentication,
       networkError,
+      server,
     },
     actions: {
       logout,
       setNetworkError,
+      setServer,
     },
   }
 
@@ -105,7 +110,7 @@ export default function AuthContextProvider(props) {
     <AuthContext.Provider value={value}>
       <AuthenticationContextProvider
         config={{
-          server: BASE_URL,
+          server,
           tokenid: TOKEN_ID,
           timeout: HTTP_GET_MAX_WAIT_TIME,
         }}
