+++ src/context/StoreContext.js
@@ -7,7 +7,6 @@ import PropTypes from 'prop-types'
 import useLocalStorage from '@hooks/useLocalStorage'
 import * as useULS from '@hooks/useUserLocalStorage'
 import { AuthContext } from '@context/AuthContext'
-import { BASE_URL } from '@common/constants'
 
 export const StoreContext = createContext({})
 
@@ -16,10 +15,12 @@ export default function StoreContextProvider(props) {
     state: {
       authentication,
       networkError: tokenNetworkError,
+      server,
     },
     actions: {
       logout,
       setNetworkError: setTokenNetworkError,
+      setServer,
     },
   } = useContext(AuthContext)
   const username = authentication?.user?.username || ''
@@ -45,7 +46,6 @@ export default function StoreContextProvider(props) {
   const [selectedQuote, setQuote] = useUserLocalStorage('selectedQuote', null)
   // TODO blm: for now we use unfoldingWord for original language bibles
   const [scriptureOwner, setScriptureOwner] = useState('unfoldingWord')
-  const [server, setServer] = useUserLocalStorage('server', BASE_URL)
   const [appRef, setAppRef] = useUserLocalStorage('appRef', 'master') // default for app
   const [bibleReference, setBibleReference] = useUserLocalStorage('bibleReference', {
     bookId: 'mat',
